Moved multiple auxiliary functions into a central file, added descriptions and function imports to CLI
