+++ Functions/cosh.py
@@ -1,5 +1,8 @@
-from Power_Function import power_function
-from Constants import get_e
+# Goal: Calculate the hyperbolic cosine of a given variable x
+# Author: Wayne
+
+from power_function import power_function
+from constants import get_e
 
 # Goal: calculate cosh(x)
 # Method: exponential, cosh(x) = (e^x + e^(-x))/2
