+++ Functions/power_function.py
@@ -1,5 +1,12 @@
 #Goal: Calculate the specified power of a specified base
 #Handling: Does not compute complex numbers 
+#Author: Amina
+
+from auxiliary_functions import floor_function
+from auxiliary_functions import sqrt_function
+from auxiliary_functions import rounding_function
+from auxiliary_functions import absolute_value
+from auxiliary_functions import decimal_to_fraction
 
 def power_function (args):
     base = 0
@@ -107,84 +114,3 @@ def power_fractional (base, power):
         return -0.0
 
     return decimal_result * integer_result
-
-def floor_function(power):
-    power *= 1.0
-    number = str(power)
-
-    #Case 1: Power is a non-integer negative number
-    if (number[0] == "-" and power != int(power)):
-        number = int(power) - 1 
-    #Case 2: Power is positive rational number
-    else: 
-        number = int(power)
-    
-    return number
-
-def sqrt_function(base):
-    precision = 0.0000000000001
-    previous, mid = 0, float(base)
-
-    #Uses binary search to find the closest approximation of the square root
-    while absolute_value(mid - previous) > precision:
-        previous, mid = mid, (mid + (base / mid)) / 2.0
-    
-    return mid
-
-def rounding_function(number):
-    #Rounds the result to 11 digits after the decimals
-    precision = 0.00000000001
-    result = str(number)
-    first_half = result[ : result.find(".")]
-    second_half = result[result.find(".") : ]
-
-    if (len(second_half) < 13):
-        return number
-
-    precision_factor = second_half[1 : 13]
-    second_half = second_half[ : 12]
-
-    if (int(precision_factor[11]) < 5):
-        full = float(first_half + second_half)
-    else:
-        value = float(second_half) + precision
-        if (value == 1):
-            full = floor_function(number)
-        else:
-            full = float(first_half) + (float(second_half) + precision)
-
-    return full
-
-def absolute_value(number):
-    if (number < 0):
-        number *= -1
-
-    return number
-
-def decimal_to_fraction(decimal):
-    precision = 0.0000000000001
-    number = int(floor_function(decimal))
-    decimal -= number
-    if (decimal < precision):
-        return (number, 1)
-    elif (1 - precision < decimal):
-        return (number + 1, 1)
-
-    low_numerator = 0
-    low_denominator = 1
-    high_numerator = 1
-    high_denominator = 1
-
-    #Uses binary search to find the closest fraction to the decimal
-    while True:
-        middle_numerator = low_numerator + high_numerator
-        middle_denominator = low_denominator + high_denominator
-
-        if middle_denominator * (decimal + precision) < middle_numerator:
-            high_numerator = middle_numerator
-            high_denominator = middle_denominator
-        elif middle_numerator < (decimal - precision) * middle_denominator:
-            low_numerator = middle_numerator
-            low_denominator = middle_denominator
-        else:
-            return (number * middle_denominator + middle_numerator, middle_denominator)
